requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
groq==0.18.0
python-dotenv==1.0.0
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
anyio==3.7.1
typing-extensions>=4.10,<5